refactor: timetableStore의 type 분리
